moved util functions
